+++ src/js/components/lightbox.js
@@ -1,4 +1,5 @@
 import Slideshow from '../mixin/slideshow';
+import Animations from './internal/lightbox-animations';
 
 function plugin(UIkit) {
 
@@ -9,7 +10,7 @@ function plugin(UIkit) {
     UIkit.use(Slideshow);
 
     var {mixin, util} = UIkit;
-    var {$, $$, addClass, Animation, ajax, append, assign, attr, css, doc, docEl, data, getImage, hasAttr, html, index, on, pointerDown, pointerMove, removeClass, Transition, trigger} = util;
+    var {$, $$, addClass, ajax, append, assign, attr, css, doc, docEl, data, getImage, html, index, on, pointerDown, pointerMove, removeClass, Transition, trigger} = util;
 
     UIkit.component('lightbox', {
 
@@ -18,7 +19,8 @@ function plugin(UIkit) {
         props: {
             animation: String,
             toggle: String,
-            autoplay: Number,
+            autoplay: Boolean,
+            autoplayInterval: Number,
             videoAutoplay: Boolean
         },
 
@@ -91,10 +93,7 @@ function plugin(UIkit) {
         methods: {
 
             _init() {
-                return this.panel = this.panel || UIkit.lightboxPanel({
-                    autoplay: this.autoplay,
-                    videoAutoplay: this.videoAutoplay,
-                    animation: this.animation,
+                return this.panel = this.panel || UIkit.lightboxPanel(assign({}, this.$props, {
                     items: this.toggles.reduce((items, el) => {
                         items.push(['href', 'caption', 'type', 'poster'].reduce((obj, attr) => {
                             obj[attr === 'href' ? 'source' : attr] = data(el, attr);
@@ -102,7 +101,7 @@ function plugin(UIkit) {
                         }, {}));
                         return items;
                     }, [])
-                });
+                }));
             },
 
             show(index) {
@@ -138,16 +137,20 @@ function plugin(UIkit) {
             items: [],
             cls: 'uk-open',
             clsPage: 'uk-lightbox-page',
-            clsItem: 'uk-lightbox-item',
+            selList: '.uk-lightbox-items',
             attrItem: 'uk-lightbox-item',
+            initialAnimation: 'scale',
+            pauseOnHover: false,
+            velocity: 2,
+            Animations: Animations(UIkit),
             template: `<div class="uk-lightbox uk-overflow-hidden">
                             <ul class="uk-lightbox-items"></ul>
-                            <div class="uk-lightbox-toolbar uk-position-top uk-text-right">
+                            <div class="uk-lightbox-toolbar uk-position-top uk-text-right uk-transition-slide-top uk-transition-opaque">
                                 <button class="uk-lightbox-toolbar-icon uk-close-large" type="button" uk-close uk-toggle="!.uk-lightbox"></button>
                              </div>
-                            <a class="uk-lightbox-button uk-position-center-left uk-position-medium" href="#" uk-slidenav-previous uk-lightbox-item="previous"></a>
-                            <a class="uk-lightbox-button uk-position-center-right uk-position-medium" href="#" uk-slidenav-next uk-lightbox-item="next"></a>
-                            <div class="uk-lightbox-toolbar uk-lightbox-caption uk-position-bottom uk-text-center"></div>
+                            <a class="uk-lightbox-button uk-position-center-left uk-position-medium uk-transition-fade" href="#" uk-slidenav-previous uk-lightbox-item="previous"></a>
+                            <a class="uk-lightbox-button uk-position-center-right uk-position-medium uk-transition-fade" href="#" uk-slidenav-next uk-lightbox-item="next"></a>
+                            <div class="uk-lightbox-toolbar uk-lightbox-caption uk-position-bottom uk-text-center uk-transition-slide-bottom uk-transition-opaque"></div>
                         </div>`
         },
 
@@ -155,12 +158,9 @@ function plugin(UIkit) {
 
             this.$mount(append(this.container, this.template));
 
-            this.list = $('.uk-lightbox-items', this.$el);
-            this.toolbars = $$('.uk-lightbox-toolbar', this.$el);
-            this.nav = $$('a[uk-lightbox-item]', this.$el);
             this.caption = $('.uk-lightbox-caption', this.$el);
 
-            this.items.forEach((el, i) => append(this.list, `<li class="${this.clsItem} item-${i}"></li>`));
+            this.items.forEach((el, i) => append(this.list, `<li></li>`));
 
         },
 
@@ -181,7 +181,7 @@ function plugin(UIkit) {
                 self: true,
 
                 delegate() {
-                    return `.${this.clsItem}`;
+                    return `${this.selList} > *`;
                 },
 
                 handler(e) {
@@ -198,9 +198,7 @@ function plugin(UIkit) {
                 self: true,
 
                 handler() {
-
                     addClass(docEl, this.clsPage);
-
                 }
             },
 
@@ -210,14 +208,7 @@ function plugin(UIkit) {
 
                 self: true,
 
-                handler() {
-
-                    addClass(this.caption, 'uk-animation-slide-bottom');
-                    attr(this.toolbars, 'hidden', '');
-                    attr(this.nav, 'hidden', '');
-                    this.showControls();
-
-                }
+                handler: 'showControls'
             },
 
             {
@@ -226,13 +217,7 @@ function plugin(UIkit) {
 
                 self: true,
 
-                handler() {
-
-                    removeClass(this.caption, 'uk-animation-slide-bottom');
-                    attr(this.toolbars, 'hidden', '');
-                    attr(this.nav, 'hidden', '');
-
-                }
+                handler: 'hideControls'
             },
 
             {
@@ -242,9 +227,7 @@ function plugin(UIkit) {
                 self: true,
 
                 handler() {
-
                     removeClass(docEl, this.clsPage);
-
                 }
             },
 
@@ -307,15 +290,16 @@ function plugin(UIkit) {
 
                 self: true,
 
-                handler() {
+                handler(e, _, el) {
 
-                    var caption = this.getItem().caption;
-                    css(this.caption, 'display', caption ? '' : 'none');
-                    html(this.caption, caption);
+                    var i = index(el),
+                        caption = this.getItem(i).caption;
+                        css(this.caption, 'display', caption ? '' : 'none');
+                        html(this.caption, caption);
 
-                    for (var i = 0; i <= this.preload; i++) {
-                        this.loadItem(this.getIndex(this.index + i));
-                        this.loadItem(this.getIndex(this.index - i));
+                    for (var j = 0; j <= this.preload; j++) {
+                        this.loadItem(this.getIndex(i + j));
+                        this.loadItem(this.getIndex(i - j));
                     }
 
                 }
@@ -450,44 +434,17 @@ function plugin(UIkit) {
                 clearTimeout(this.controlsTimer);
                 this.controlsTimer = setTimeout(this.hideControls, this.delayControls);
 
-                if (!hasAttr(this.toolbars, 'hidden')) {
-                    return;
-                }
-
-                animate(this.toolbars[0], 'uk-animation-slide-top');
-                animate(this.toolbars[1], 'uk-animation-slide-bottom');
-
-                attr(this.nav, 'hidden', this.items.length <= 1 ? '' : null);
-
-                if (this.items.length > 1) {
-                    animate(this.nav, 'uk-animation-fade');
-                }
+                addClass(this.$el, 'uk-active uk-transition-active');
 
             },
 
             hideControls() {
-
-                if (hasAttr(this.toolbars, 'hidden')) {
-                    return;
-                }
-
-                animate(this.toolbars[0], 'uk-animation-slide-top', 'out');
-                animate(this.toolbars[1], 'uk-animation-slide-bottom', 'out');
-
-                if (this.items.length > 1) {
-                    animate(this.nav, 'uk-animation-fade', 'out');
-                }
-
+                removeClass(this.$el, 'uk-active uk-transition-active');
             }
 
         }
 
     });
-
-    function animate(el, animation, dir = 'in') {
-        attr(el, 'hidden', null);
-        Animation[dir](el, animation).then(() => dir === 'out' && attr(el, 'hidden', ''));
-    }
 
     function getIframe(src, width, height, autoplay) {
         return `<iframe src="${src}" width="${width}" height="${height}" style="max-width: 100%; box-sizing: border-box;" frameborder="0" allowfullscreen uk-video="autoplay: ${autoplay}" uk-responsive></iframe>`;