+++ src/js/mixin/slideshow.js
@@ -1,42 +1,60 @@
+import Animations from './internal/slideshow-animations';
+
 function plugin(UIkit) {
 
     if (plugin.installed) {
         return;
     }
 
-    var {addClass, css, filter, hasClass, removeClass, toggleClass, $, attr, doc, fastdom, getIndex, noop, off, on, pointerDown, pointerMove, pointerUp, preventClick, Promise, requestAnimationFrame, Transition, trigger} = UIkit.util;
+    var {$$, $, addClass, assign, attr, css, doc, endsWith, fastdom, getIndex, hasClass, index, noop, off, on, pointerDown, pointerMove, pointerUp, preventClick, Promise, removeClass, toggleClass, Transition, trigger} = UIkit.util;
 
     UIkit.mixin.slideshow = {
 
         attrs: true,
 
         props: {
-            autoplay: Number,
+            autoplay: Boolean,
+            autoplayInterval: Number,
+            pauseOnHover: Boolean,
             animation: String,
-            transition: String,
-            duration: Number
+            easing: String,
+            velocity: Number
         },
 
         defaults: {
-            autoplay: 0,
+            autoplay: false,
+            autoplayInterval: 7000,
+            pauseOnHover: true,
             animation: 'slide',
-            transition: 'linear',
-            duration: 400,
+            easing: 'ease',
+            velocity: 1,
             index: 0,
             stack: [],
             threshold: 10,
             percent: 0,
-            clsActive: 'uk-active'
+            clsActive: 'uk-active',
+            clsActivated: 'uk-transition-active',
+            forwardDuration: 150,
+            initialAnimation: false,
+            Animations: Animations(UIkit)
         },
 
         computed: {
 
-            slides({clsItem}) {
-                return filter(this.list.children, `.${clsItem}`);
+            list({selList}, $el) {
+                return $(selList, $el);
             },
 
-            forwardDuration({duration}) {
-                return duration / 4;
+            slides() {
+                return $$(this.list.children);
+            },
+
+            animation({animation, Animations}) {
+                return assign(animation in Animations ? Animations[animation] : Animations.slide, {name: animation});
+            },
+
+            duration({velocity}, $el) {
+                return speedUp($el.offsetWidth / velocity);
             }
 
         },
@@ -57,6 +75,24 @@ function plugin(UIkit) {
         connected() {
             this.startAutoplay();
         },
+
+        disconnected() {
+            this.stopAutoplay();
+        },
+
+        update: [
+
+            {
+
+                read() {
+                    delete this._computeds.duration;
+                },
+
+                events: ['load', 'resize']
+
+            }
+
+        ],
 
         events: [
 
@@ -81,7 +117,7 @@ function plugin(UIkit) {
                 name: pointerDown,
 
                 delegate() {
-                    return `.${this.clsItem}`;
+                    return `${this.selList} > *`;
                 },
 
                 handler: 'start'
@@ -123,6 +159,67 @@ function plugin(UIkit) {
 
             },
 
+            {
+
+                name: 'beforeitemshow',
+
+                self: true,
+
+                handler(e, _, el) {
+                    addClass(el, this.clsActive);
+                }
+
+            },
+
+            {
+
+                name: 'itemshown',
+
+                self: true,
+
+                handler(e, _, el) {
+                    addClass(el, this.clsActivated);
+                }
+
+            },
+
+            {
+
+                name: 'itemshow itemhide',
+
+                self: true,
+
+                handler({type}, _, el) {
+                    toggleClass($$(`[${this.attrItem}="${index(el)}"]`, this.$el), this.clsActive, endsWith(type, 'show'));
+                }
+
+            },
+
+            {
+
+                name: 'itemhidden',
+
+                self: true,
+
+                handler(e, _, el) {
+                    removeClass(el, this.clsActive);
+                    removeClass(el, this.clsActivated);
+                }
+
+            },
+
+            {
+
+                name: 'itemshow itemhide itemshown itemhidden',
+
+                self: true,
+
+                handler(e, _, el) {
+                    UIkit.update(null, el);
+                }
+
+            }
+
         ],
 
         methods: {
@@ -135,32 +232,33 @@ function plugin(UIkit) {
 
                 e.preventDefault();
 
+                if (this._animation && this._animation.animation !== this.animation) {
+                    return;
+                }
+
                 var percent = 0;
                 if (this.stack.length) {
 
-                    this.percent = this._animation.percent();
+                    var {dir, percent: getPercent, cancel, translate} = this._animation;
 
-                    var dir = this._animation.dir;
-                    percent = this.percent * dir;
+                    percent = getPercent() * dir;
+
+                    this.percent = Math.abs(percent) * -dir;
 
                     this.stack.splice(0, this.stack.length);
 
-                    this._animation.cancel();
-                    this._animation.translate(Math.abs(percent));
+                    cancel();
+                    translate(Math.abs(percent));
 
                     this.index = this.getIndex(this.index - dir);
                     this.touching = true;
+
                 }
 
                 on(doc, pointerMove, this.move, true);
                 on(doc, pointerUp, this.end, true);
 
-                var el = this.slides[this.index];
-
-                this.touch = {
-                    el,
-                    start: this.pos + (percent ? el.offsetWidth * percent : 0)
-                }
+                this.touch = this.pos + this.$el.offsetWidth * percent;
 
             },
 
@@ -168,21 +266,19 @@ function plugin(UIkit) {
 
                 e.preventDefault();
 
-                var {start, el} = this.touch;
-
-                if (this.pos === this.prevPos || (!this.touching && Math.abs(start - this.pos) < this.threshold)) {
+                if (this.pos === this.prevPos || (!this.touching && Math.abs(this.touch - this.pos) < this.threshold)) {
                     return;
                 }
 
                 this.touching = true;
 
-                var percent = (this.pos - start) / el.offsetWidth;
+                var percent = (this.pos - this.touch) / this.$el.offsetWidth;
 
                 if (this.percent === percent) {
                     return;
                 }
 
-                var changed = trunc(this.percent) !== trunc(percent),
+                var prevIndex = this.getIndex(this.index - trunc(this.percent)),
                     index = this.getIndex(this.index - trunc(percent)),
                     current = this.slides[index],
                     dir = percent < 0 ? 1 : -1,
@@ -191,11 +287,13 @@ function plugin(UIkit) {
 
                 this.slides.forEach((el, i) => toggleClass(el, this.clsActive, i === index || i === nextIndex));
 
-                if (changed && this._animation) {
-                    this._animation.reset();
+                if (index !== prevIndex) {
+                    this._animation && this._animation.reset();
+                    trigger(this.$el, 'itemhide', [this, this.slides[prevIndex]]);
+                    trigger(this.$el, 'itemshow', [this, current]);
                 }
 
-                this._animation = new Transitioner(this.animation, this.transition, current, next, dir, noop);
+                this._animation = new Transitioner(this.animation, this.easing, current, next, dir, noop);
                 this._animation.translate(Math.abs(percent % 1));
 
                 this.percent = percent;
@@ -218,7 +316,7 @@ function plugin(UIkit) {
                     this.percent = Math.abs(this.percent) % 1;
                     this.index = this.getIndex(this.index - trunc(percent));
 
-                    if (this.percent < 0.2) {
+                    if (this.percent < .1) {
                         this.index = this.getIndex(percent > 0 ? 'previous' : 'next');
                         this.percent = 1 - this.percent;
                         percent *= -1;
@@ -256,64 +354,78 @@ function plugin(UIkit) {
                     return;
                 }
 
-                var hasPrev = hasClass(this.slides, 'uk-active'),
-                    dir = index === 'next'
-                            ? 1
-                            : index === 'previous'
-                                ? -1
-                                : index < this.index
-                                    ? -1
-                                    : 1;
+                var prevIndex = this.index,
+                    nextIndex = this.getIndex(index),
+                    prev = hasClass(this.slides, 'uk-active') && this.slides[prevIndex],
+                    next = this.slides[nextIndex];
 
-                index = this.getIndex(index);
-
-                if (hasPrev && index === this.index) {
+                if (prev === next) {
                     this.stack[force ? 'shift' : 'pop']();
                     return;
                 }
 
-                var prev = hasPrev && this.slides[this.index],
-                    next = this.slides[index];
-
-                trigger(this.$el, 'beforeitemshow', [this, next]);
                 prev && trigger(this.$el, 'beforeitemhide', [this, prev]);
+                trigger(this.$el, 'beforeitemshow', [this, next]);
 
-                this.index = index;
+                this.index = nextIndex;
 
-                addClass(next, this.clsActive);
+                var done = () => {
 
-                this._animation = new Transitioner(!prev ? 'scale' : this.animation, this.transition, prev || next, next, dir, () => {
-
-                    prev && removeClass(prev, this.clsActive);
+                    prev && trigger(this.$el, 'itemhidden', [this, prev]);
+                    trigger(this.$el, 'itemshown', [this, next]);
 
                     this.stack.shift();
                     if (this.stack.length) {
-                        requestAnimationFrame(() => this.show(this.stack.shift(), true));
+                        this.show(this.stack.shift(), true);
                     } else {
                         this._animation = null;
                     }
+                };
 
-                    trigger(this.$el, 'itemshown', [this, next]);
-                    UIkit.update(null, next);
+                if (prev || this.initialAnimation) {
 
-                    if (prev) {
-                        trigger(this.$el, 'itemhidden', [this, prev]);
-                        UIkit.update(null, prev);
-                    }
+                    this._show(
+                        !prev ? this.Animations[this.initialAnimation] : this.animation,
+                        force ? 'cubic-bezier(0.165, 0.840, 0.440, 1.000)' : this.easing,
+                        prev || next,
+                        next,
+                        getDirection(index, prevIndex),
+                        this.stack.length > 1,
+                        done
+                    );
 
-                });
+                } else {
 
-                this._animation.show(this.stack.length > 1 ? this.forwardDuration : this.duration, this.percent);
+                    done();
 
-                trigger(this.$el, 'itemshow', [this, next]);
-
-                if (prev) {
-                    trigger(this.$el, 'itemhide', [this, prev]);
-                    UIkit.update(null, prev);
                 }
 
-                UIkit.update(null, next);
+                prev && trigger(this.$el, 'itemhide', [this, prev]);
+                trigger(this.$el, 'itemshow', [this, next]);
+
                 fastdom.flush(); // iOS 10+ will honor the video.play only if called from a gesture handler
+
+            },
+
+            _show(animation, easing, prev, next, dir, forward, done) {
+
+                this._animation = new Transitioner(
+                    animation,
+                    easing,
+                    prev,
+                    next,
+                    dir,
+                    done
+                );
+
+                this._animation.show(prev === next
+                    ? 300
+                    : forward
+                        ? this.forwardDuration
+                        : this.duration
+                    , this.percent, forward
+                );
+
             },
 
             getIndex(index = this.index) {
@@ -325,7 +437,7 @@ function plugin(UIkit) {
                 this.stopAutoplay();
 
                 if (this.autoplay) {
-                    this.interval = setInterval(() => !this.isHovering && this.show('next'), this.autoplay);
+                    this.interval = setInterval(() => (!this.isHovering || ! this.pauseOnHover) && this.show('next'), this.autoplayInterval);
                 }
 
             },
@@ -340,149 +452,28 @@ function plugin(UIkit) {
 
     };
 
-    var diff = 0.2;
-    var Animations = {
+    function Transitioner(animation, easing, current, next, dir, cb) {
 
-        fade: {
-
-            show() {
-                return [
-                    {opacity: 0},
-                    {opacity: 1}
-                ];
-            },
-
-            percent(current) {
-                return 1 - css(current, 'opacity');
-            },
-
-            translate(percent) {
-                return [
-                    {opacity: 1 - percent},
-                    {opacity: percent}
-                ];
-            }
-
-        },
-
-        slide: {
-
-            show(dir) {
-                return [
-                    {transform: `translate3d(${dir * -100}%, 0, 0)`},
-                    {transform: 'translate3d(0, 0, 0)'}
-                ];
-            },
-
-            percent(current) {
-                return Math.abs(css(current, 'transform').split(',')[4] / current.offsetWidth);
-            },
-
-            translate(percent, dir) {
-                return [
-                    {transform: `translate3d(${dir * -100 * percent}%, 0, 0)`},
-                    {transform: `translate3d(${dir * 100 * (1 - percent)}%, 0, 0)`}
-                ];
-            }
-
-        },
-
-        scale: {
-
-            show() {
-                return [
-                    {opacity: 0, transform: `scale3d(${1 - diff}, ${1 - diff}, 1)`},
-                    {opacity: 1, transform: 'scale3d(1, 1, 1)'}
-                ];
-            },
-
-            percent(current) {
-                return 1 - css(current, 'opacity');
-            },
-
-            translate(percent) {
-                var scale1 = 1 - diff * percent,
-                    scale2 = 1 - diff + diff * percent;
-
-                return [
-                    {opacity: 1 - percent, transform: `scale3d(${scale1}, ${scale1}, 1)`},
-                    {opacity: percent, transform: `scale3d(${scale2}, ${scale2}, 1)`}
-                ];
-            }
-
-        },
-
-        swipe: {
-
-            show(dir) {
-
-                if (dir < 0) {
-                    return [
-                        {opacity: 1, transform: `translate3d(100%, 0, 0)`, zIndex: 0},
-                        {opacity: 1, transform: `scale3d(1, 1, 1) translate3d(0, 0, 0)`, zIndex: -1},
-                    ];
-                } else {
-                    return [
-                        {opacity: 0.3, transform: `scale3d(${1 - diff}, ${1 - diff}, 1) translate3d(-20%, 0, 0)`, zIndex: -1},
-                        {opacity: 1, transform: 'translate3d(0, 0, 0)', zIndex: 0}
-                    ];
-                }
-
-
-            },
-
-            percent(current, next, dir) {
-
-                var el = dir < 0 ? current : next,
-                    percent = Math.abs(css(el, 'transform').split(',')[4] / el.offsetWidth);
-
-                return dir < 0 ? percent : 1 - percent;
-            },
-
-            translate(percent, dir) {
-                var scale;
-
-                if (dir < 0) {
-                    scale = 1 - diff * (1 - percent);
-                    return [
-                        {opacity: 1, transform: `translate3d(${100 * percent}%, 0, 0)`, zIndex: 0},
-                        {opacity: 0.3 + 0.7 * percent, transform: `scale3d(${scale}, ${scale}, 1) translate3d(${-20 * (1 - percent)}%, 0, 0)`, zIndex: -1},
-                    ];
-                } else {
-                    scale = 1 - diff * percent;
-                    return [
-                        {opacity: 1 - 0.7 * percent, transform: `scale3d(${scale}, ${scale}, 1) translate3d(${-20 * percent}%, 0, 0)`, zIndex: -1},
-                        {opacity: 1, transform: `translate3d(${100 * (1 - percent)}%, 0, 0)`, zIndex: 0}
-                    ];
-                }
-
-            }
-
-        },
-
-    };
-
-    function Transitioner(animation, transition, current, next, dir, cb) {
-
-        animation = animation in Animations ? Animations[animation] : Animations.slide;
-
-        var props = animation.show(dir);
+        var {percent, translate, show} = animation;
+        var props = show(dir);
 
         return {
 
+            animation,
             dir,
             current,
             next,
 
-            show(duration, percent = 0) {
+            show(duration, percent = 0, linear) {
 
+                var ease = linear ? 'linear' : easing;
                 duration -= Math.round(duration * percent);
 
                 this.translate(percent);
 
                 return Promise.all([
-                    Transition.start(current, props[0], duration, transition),
-                    Transition.start(next, props[1], duration, transition)
+                    Transition.start(current, props[0], duration, ease),
+                    Transition.start(next, props[1], duration, ease)
                 ]).then(() => {
                     this.reset();
                     cb();
@@ -490,46 +481,37 @@ function plugin(UIkit) {
             },
 
             stop() {
-                return Promise.all([
-                    Transition.stop(next),
-                    Transition.stop(current)
-                ]);
+                return Transition.stop([next, current]);
             },
 
             cancel() {
-                return Promise.all([
-                    Transition.cancel(next),
-                    Transition.cancel(current)
-                ]);
+                Transition.cancel([next, current]);
             },
 
             reset() {
                 for (var prop in props[0]) {
-                    css([next, current[0]], prop, '');
+                    css([next, current], prop, '');
                 }
             },
 
             forward(duration) {
 
                 var percent = this.percent();
-
-                return Promise.all([
-                    Transition.cancel(next),
-                    Transition.cancel(current)
-                ]).then(() => this.show(duration, percent));
+                Transition.cancel([next, current]);
+                this.show(duration, percent, true);
 
             },
 
             translate(percent) {
 
-                var props = animation.translate(percent, dir);
+                var props = translate(percent, dir);
                 css(current, props[0]);
                 css(next, props[1]);
 
             },
 
             percent() {
-                return animation.percent(current, next, dir);
+                return percent(current, next, dir);
             }
 
         }
@@ -539,6 +521,20 @@ function plugin(UIkit) {
     // polyfill for Math.trunc (IE)
     function trunc(x) {
         return ~~x;
+    }
+
+    function getDirection(index, prevIndex) {
+        return index === 'next'
+            ? 1
+            : index === 'previous'
+                ? -1
+                : index < prevIndex
+                    ? -1
+                    : 1;
+    }
+
+    function speedUp(x) {
+        return .0001785714 * x * x + .3214285714 * x + 342.8571428571; // parabel through (400,500; 600,600; 1800,1500)
     }
 
 }